requirements.txt
python-dotenv==1.0.1
pytest>=7.0.0
black==23.11.0
isort==5.12.0
google-api-python-client==2.108.0
requests==2.31.0
google-generativeai==0.3.2
youtube-transcript-api==0.6.2
